ecomm/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
sentence-transformers>=2.2.0
pinecone>=2.2.0
torch>=1.13.0
transformers>=4.20.0
python-dotenv>=1.0.0 